requirements.txt
requests
pillow
python-telegram-bot